Selector-based cropping for screenshots

+++ src/tests/saveScreenshot.js
@@ -1,8 +1,10 @@
 const { validate } = require("doc-detective-common");
 const { log } = require("../utils");
+const os = require("os");
 const path = require("path");
 const fs = require("fs");
 const PNG = require("pngjs").PNG;
+const sharp = require("sharp");
 const pixelmatch = require("pixelmatch");
 
 exports.saveScreenshot = saveScreenshot;
@@ -24,7 +26,9 @@ async function saveScreenshot(config, step, driver) {
   // Set file name
   if (!step.path) {
     step.path = `${step.id}.png`;
-    if (step.directory) { step.path = path.join(step.directory, step.path); }
+    if (step.directory) {
+      step.path = path.join(step.directory, step.path);
+    }
   }
   let filePath = step.path;
 
@@ -34,7 +38,7 @@ async function saveScreenshot(config, step, driver) {
   if (!fs.existsSync(dir)) {
     fs.mkdirSync(dir, { recursive: true });
   }
-    
+
   // Check if file already exists
   let existFilePath;
   if (fs.existsSync(filePath)) {
@@ -70,6 +74,79 @@ async function saveScreenshot(config, step, driver) {
     result.status = "FAIL";
     result.description = `Couldn't save screenshot. ${error}`;
     return result;
+  }
+
+  // If crop is set, found bounds of element and crop image
+  if (step.crop) {
+    let padding = { top: 0, right: 0, bottom: 0, left: 0 };
+    if (typeof step.crop.padding === "number") {
+      padding.top = step.crop.padding;
+      padding.right = step.crop.padding;
+      padding.bottom = step.crop.padding;
+      padding.left = step.crop.padding;
+    } else if (typeof step.crop.padding === "object") {
+      padding = step.crop.padding;
+    }
+
+    // Get pixel density
+    const pixelDensity = await driver.execute(() => window.devicePixelRatio);
+
+    // Get the element using the provided selector
+    const element = await driver.$(step.crop.selector);
+
+    // Get the bounding rectangle of the element
+    const rect = {
+      ...(await element.getSize()),
+      ...(await element.getLocation()),
+    };
+    log(config, "debug", { rect });
+
+    // Calculate the padding based on the provided padding values
+    rect.x -= padding.left;
+    rect.y -= padding.top;
+    rect.width += padding.left + padding.right;
+    rect.height += padding.top + padding.bottom;
+
+    // Scale the values based on the pixel density
+    rect.x *= pixelDensity;
+    rect.y *= pixelDensity;
+    rect.width *= pixelDensity;
+    rect.height *= pixelDensity;
+
+    // Round the values to integers
+    rect.x = Math.round(rect.x);
+    rect.y = Math.round(rect.y);
+    rect.width = Math.round(rect.width);
+    rect.height = Math.round(rect.height);
+
+    log(config, "debug", { padded_rect: rect });
+
+    // TODO: Add error handling for out of bounds
+
+    // Create a new PNG object with the dimensions of the cropped area
+    const croppedPath = path.join(dir, "cropped.png");
+    sharp(filePath)
+      .extract({
+        left: rect.x,
+        top: rect.y,
+        width: rect.width,
+        height: rect.height,
+      })
+      .toFile(croppedPath, (err, info) => {
+        if (err) {
+          result.status = "FAIL";
+          result.description = `Couldn't crop image. ${err}`;
+          return result;
+        }
+      });
+    
+      // Wait for the file to be written
+    while (!fs.existsSync(croppedPath)) {
+      await new Promise((resolve) => setTimeout(resolve, 100));
+    }
+
+    // Replace the original file with the cropped file
+    fs.renameSync(croppedPath, filePath);
   }
 
   // If file already exists
@@ -110,8 +187,6 @@ async function saveScreenshot(config, step, driver) {
       if (percentDiff > step.maxVariation) {
         if (step.overwrite == "byVariation") {
           // Replace old file with new file
-          // TODO: Not working
-          fs.unlinkSync(existFilePath);
           fs.renameSync(filePath, existFilePath);
         }
         result.status = "FAIL";
@@ -123,13 +198,15 @@ async function saveScreenshot(config, step, driver) {
         result.description = `Screenshots are within maximum accepted variation: ${percentDiff.toFixed(
           2
         )}%.`;
-        fs.unlinkSync(filePath);
+        if (step.overwrite != "true") {
+          fs.unlinkSync(filePath);
+        }
       }
     }
 
     if (step.overwrite == "true") {
       // Replace old file with new file
-      fs.unlinkSync(existFilePath);
+      result.description += ` Overwrote existing file.`;
       fs.renameSync(filePath, existFilePath);
     }
   }
